+++ monitor/types.hpp
@@ -44,7 +44,7 @@ using SensorDefinition = std::tuple<std::string,
                                     bool,
                                     std::string,
                                     size_t,
-                                    size_t>;
+                                    int64_t>;
 
 constexpr auto fanNameField = 0;
 constexpr auto funcDelay = 1;
